requirements.txt
python-dotenv==1.1.0
flask==3.1.1
Flask-Cors==6.0.0
flask_sqlalchemy==3.1.1
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
pymysql==1.1.1
cryptography==45.0.2